Fix clash with WebUI's default XYZ script
The default WebUI comes with a X/Y/Z plotting script. It's called "X/Y/Z plot" and has the filename `xyz_grid.py` as just like this one. I think it would be a good idea to rename the file and also rename the script title to "X/Y/Z HTML plot."